Add killProcess helper for test functions.
This function is useful when a process is being killed and then restarted, because if a process has not been fully killed when its restart is attempted it will silently fail to restart. This function ensures a process has been terminated before returning because  will error if given a pid for a process that doesn't exist.